changed service and added service implementation
