+++ src/main/java/com/example/crud1/usersLeave/UsersLeaveService.java
@@ -6,28 +6,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class UsersLeaveService {
+public interface UsersLeaveService {
+
+    List<UsersLeaveEntity> getAllUsers();
+
+    List<UsersLeaveEntity> findById(String userName);
+
+    UsersLeaveEntity saveUsersDetails(UsersLeaveEntity record);
     
-    @Autowired
-    private UsersLeaveRepo usersLeaveRepo;
-
-    public UsersLeaveService(UsersLeaveRepo usersLeaveRepo) {
-        super();
-        this.usersLeaveRepo = usersLeaveRepo;
-    }
-
-    public List<UsersLeaveEntity> getAllUsers() {
-        return usersLeaveRepo.findAll();
-    }
-
-    // public UsersLeaveEntity findById(userName) {
-    //     return usersLeaveRepo.findByUserName();
-    // }
-
-    public List<UsersLeaveEntity> findById(String userName) {
-        return usersLeaveRepo.findByUserName(userName);
-    }
-
-    public UsersLeaveEntity saveUsersDetails(UsersLeaveEntity record) {
-        return usersLeaveRepo.save(record);
-    }}
+   }
